Move types to lib
