+++ lib/array.js
@@ -1,5 +1,5 @@
 'use strict'
-var util = require('../util')
+var util = require('./util')
 
 module.exports = function (length, itemType) {
   if (typeof length !== 'number') throw new TypeError('length must be a number')
