+++ lib/bound.js
@@ -1,5 +1,5 @@
 'use strict'
-var util = require('../util')
+var util = require('./util')
 
 module.exports = function (itemType, checkValue) {
   if (!util.isAbstractCodec(itemType)) throw new TypeError('itemType is invalid codec')
