+++ lib/index.js
@@ -1,8 +1,8 @@
 'use strict'
-module.exports = exports = require('./types/object')
+module.exports = exports = require('./object')
 
 // numbers
-var numbers = require('./types/numbers')
+var numbers = require('./numbers')
 exports.Byte = numbers.Byte
 exports.Int8 = numbers.Int8
 exports.UInt8 = numbers.UInt8
@@ -24,17 +24,17 @@ exports.DoubleBE = numbers.DoubleBE
 exports.DoubleLE = numbers.DoubleLE
 
 // array & vararray & sequence
-exports.Array = require('./types/array')
-exports.VarArray = require('./types/vararray')
-exports.Sequence = require('./types/sequence')
+exports.Array = require('./array')
+exports.VarArray = require('./vararray')
+exports.Sequence = require('./sequence')
 
 // buffer & varbuffer
-exports.Buffer = require('./types/buffer')
-exports.VarBuffer = require('./types/varbuffer')
+exports.Buffer = require('./buffer')
+exports.VarBuffer = require('./varbuffer')
 
 // string & varstring
-exports.String = require('./types/string')
-exports.VarString = require('./types/varstring')
+exports.String = require('./string')
+exports.VarString = require('./varstring')
 
 // bound
-exports.Bound = require('./types/bound')
+exports.Bound = require('./bound')
