+++ lib/numbers.js
