+++ lib/object.js
@@ -1,5 +1,5 @@
 'use strict'
-var util = require('../util')
+var util = require('./util')
 
 module.exports = function (items) {
   if (!Array.isArray(items)) throw new TypeError('items must be an Array instance')
