+++ lib/sequence.js
@@ -1,5 +1,5 @@
 'use strict'
-var util = require('../util')
+var util = require('./util')
 
 module.exports = function (types) {
   if (!Array.isArray(types)) throw new TypeError('types must be an Array instance')
