+++ lib/string.js
