+++ lib/varbuffer.js
@@ -1,5 +1,5 @@
 'use strict'
-var util = require('../util')
+var util = require('./util')
 
 module.exports = function (lengthType) {
   if (!util.isAbstractCodec(lengthType)) throw new TypeError('lengthType is invalid codec')
