+++ lib/varstring.js
@@ -1,6 +1,6 @@
 'use strict'
 var VarBuffer = require('./varbuffer')
-var util = require('../util')
+var util = require('./util')
 
 module.exports = function (lengthType, encoding) {
   if (!util.isAbstractCodec(lengthType)) throw new TypeError('lengthType is invalid codec')
