+++ test/varbuffer.js
@@ -1,6 +1,6 @@
 'use strict'
 var test = require('tape').test
-var varstruct = require('../../')
+var varstruct = require('../')
 
 var varbuffer = varstruct.VarBuffer(varstruct.UInt32BE)
